backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-docx==1.1.0
PyPDF2==3.0.1
PyMuPDF==1.23.8
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pydantic==2.5.0
google-generativeai
pydantic-settings==2.0.3